Fix issues, delete Nether not uses

+++ src/main/java/cn/nukkit/level/generator/biome/NormalBiome.java
@@ -12,7 +12,7 @@ public abstract class NormalBiome extends Biome {
     }
 
     @Override
-    public int setWaterColor() {
+    public float setWaterColor() {
         return this.waterColor;
     }
 }
